feat: Add audiences and filters to notifications

+++ whistle-server/preference/serializers.py
@@ -3,18 +3,30 @@ import logging
 from django.db import transaction
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
+from sendgrid.helpers.mail import category
 
 from external_user.models import ExternalUser
 from preference.models import (
     ExternalUserPreferenceChannel,
     ExternalUserPreference,
+    ChannelChoices,
 )
 
 
 class ExternalUserPreferenceChannelSerializer(serializers.ModelSerializer):
+    slug = serializers.CharField(max_length=255)
+
     class Meta:
         model = ExternalUserPreferenceChannel
         fields = ["id", "slug", "enabled"]
+
+    def validate_slug(self, value):
+        # Convert input to uppercase
+        value_upper = value.upper()
+        # Check if the value is a valid choice
+        if value_upper not in ChannelChoices.values:
+            raise serializers.ValidationError(f"'{value}' is not a valid choice.")
+        return value_upper
 
 
 class ExternalUserPreferenceSerializer(serializers.ModelSerializer):
@@ -52,7 +64,12 @@ class ExternalUserPreferenceSerializer(serializers.ModelSerializer):
                         user_preference=user_preference, **channel
                     )
 
-                logging.info("Preference with id: %s created for user: %s in org: %s", user_preference.id, external_user.id, org.id)
+                logging.info(
+                    "Preference with id: %s created for user: %s in org: %s",
+                    user_preference.id,
+                    external_user.id,
+                    org.id,
+                )
 
                 return user_preference
         except ExternalUser.DoesNotExist:
@@ -66,7 +83,7 @@ class ExternalUserPreferenceSerializer(serializers.ModelSerializer):
                 "invalid_external_id",
             )
 
-    def update(self, instance, validated_data):
+    def update(self, instance, validated_data, **kwargs):
         org = self.context["request"].user
         external_id = self.context.get("external_id")
         with transaction.atomic():
@@ -74,14 +91,27 @@ class ExternalUserPreferenceSerializer(serializers.ModelSerializer):
             instance.slug = validated_data.get("slug", instance.slug)
             instance.save()
 
-            ExternalUserPreferenceChannel.objects.filter(
-                user_preference=instance
-            ).delete()
-            for channel_data in channels_data:
-                ExternalUserPreferenceChannel.objects.create(
-                    user_preference=instance, **channel_data
-                )
+            if self.partial:
+                for channel in channels_data:
+                    ExternalUserPreferenceChannel.objects.update_or_create(
+                        user_preference=instance,
+                        slug=channel.get("slug"),
+                        defaults={"enabled": channel.get("enabled")},
+                    )
+            else:
+                ExternalUserPreferenceChannel.objects.filter(
+                    user_preference=instance
+                ).delete()
+                for channel in channels_data:
+                    ExternalUserPreferenceChannel.objects.create(
+                        user_preference=instance, **channel
+                    )
 
-            logging.info("Preference with id: %s updated for user: %s in org: %s", instance.id, instance.user.id, org.id)
+            logging.info(
+                "Preference with id: %s updated for user: %s in org: %s",
+                instance.id,
+                instance.user.id,
+                org.id,
+            )
 
             return instance
